feat: Pay with card
This commit includes all the functionality related to payments with Stripe. It lets a user, logged in or not, create a order using a credit card.

You can test this with card: 4242 4242 4242 4242. Which is one of the ones accepted by Stripe in developer mode

It includes as well a confirmation in the backend before the order is passes that the payment has been processed correctly, as to prevent malicious attacks